Automatically create a publishing group for each new user named after their username and add them to it.
We do not currently have a ui for managing users and publishing groups.

For UAT we need admins to be able to create sites for users that only those users can edit.

This feature automatically creates a publishing group when a user record is created, giving it the username
of the user, and adds the user to it.

If the group already exists, the user is added to the existing group and no new group is created.

The artisan console command astro:adduser has been updated to make group name optional. If the user account
already exists and no group name is added it creates (if it does not exist) a group with the user's username
and adds them to it (so can be used to update existing user accounts on the test server).