Inject preview bar to stick at the top of every preview page.
A regex attempts to insert the content of the preview template html directly below the opening body tag.

The preview bar has position fixed to stick it at the top of the page, a fixed height, and an empty div with the
same height after it to push the top of the page content down below it.