Modify draft preview route so menu link and potentially other links can be modified by renderer in preview mode.
Draft preview url is now similar to published page preview url '/draft/{domain-name}/{path-to-page}' instead of '/draft/{id}'.

This makes it simpler to do a string replace on the beginning of URLs to map real published URLs to ones that can be displayed in preview mode.

Added a getter to (page) state to retrieve the preview and published urls for the current page in the editor.