Add a component that fixes the change event in Element UI's select component, then use it in our site users view.
Plus some other minor enhancements to functionality and wording.
Element UI's select component fires a change event when the value (v-model data) of the select changes, either indirectly or because a user selects/clicks an option.
This means if we modify the underlying data bound via v-model it fires a change event, but it shouldn't and led to a bug in the site users view.