Use .env vars in build, update webpack, remove hardcoded cms-prototype-blocks.
Makes the content of the .env file available in the webpack build script as process.ENV.{variable-name}

Replaces the hard-coded imports from cms-prototype-blocks in Block.vue and Preview.vue with an alias "theme"
which is aliased to whatever is set for DEFINITIONS_PATH in the .env file.

This allows defintion repositories to be named something other than "cms-prototype-blocks" and means they
no longer HAVE to be installed in node_modules as a dependency.

This should also allow us to specify other build-time variables, for example URLs, paths, etc, and thus
aid in the separation of editor from API.

As my knowledge of webpack is limited, there is the possibility that this also breaks things in ways I cannot
begin to comprehend...