Update API to return published "status" for pages.
Added an attribute getter on Page model to calculate the status of a page based on the state of its published version if one exists.

Currently not eager loaded due to restrictions / lack-of-understanding of eloquent relations.

Used the existing status options as defined in store/page.js and components/TopBar.vue:

- new = Not yet published
- draft = Published but modified
- published = Published and up to date