CreatePageModal generates slug automatically from page title unless user edits slug.
Fixes #227 by using a computed property for the page title in the CreatePageModal with a setter
that checks if the user has manually edited the slug and if not sets the slug to the slugified title.

The slug input now reacts to the change event instead of the focus event to determine whether or not
the user has started editing it manually.

Also minor refactor on the createForm data model to make it saner.