Store block errors with blocks when adding page
This already occurs when pages are updated, but checking for block errors
did not occur when pages were created (as these were from templates).

As we have removed some default values from some blocks pages containing
these blocks can now be invalid after creation.

As the editor relies on the API to report block errors when first loading pages
this results in the editor not being aware of what the block errors are until a
page has been saved at least once with the errors in it.

This fix duplicates the logic from the updatepagecontent action into the Page::saveblocks() method so that it is also used when a page is first created.