Add new route to unlink a media item from a specific site.
Create a new Permission image.unlink - remember to run php artisan astro:permissions add
to install it.

Add a new route: DELETE /api/v1/sites/{site_id}/media/{media_id}

Returns 204 (no content) if successful
Returns 404 if the media is not linked to the site
Returns 403 if user is not authorized to make the request.