Make sure confirmation dialog and router work in iframe when changing pages
The confirmation dialog was being displayed in the iframe underneath the
loading panel (so invisible) when changing pages (and unauthorised).

Also, the router would only reload the iframe which didn't work out
too well for all concerned.

This fix fixes all that by sharing the router and the vue instance that
shows the confirmation dialog.