Fix broken test
The test was failing as it was expecting the definition to have been stored in
redis, however as previous tests had already loaded that definition it is cached
(as a DefinitionObject, not the json) in memory by the fromDefinitionFile() method.

As the test appears to just want to test that the definition loads from file when not
in redis, removing the offending lines is fine.