Update composer files to laravel framework 5.4.36
This removes the dependency on @theninja fork of laravel.

@theninja forked laravel to add support for validation messages for the size rule that
were based on attribute type.

Laravel closed the pull request.

I'm not even sure if we are using these anywhere @theninja?

Astro appears to work after the update to laravel framework 5.4.36.

Any issues that may occur are likely to represent when saving pages.