Add a try-catch block to catch JWT library exceptions instead of throwing 500's
The JWT parse() method throws an exception if the token is not a syntactically valid jwt string
(for example when no token is specified at all).

This results in unhelpful 500 errors being sent to the client.

Catching the exception and just returning null for the user results in a sensible
401.