Fix an obscure bug that in rare scenarios was causing async-validator to validate fields with absolutely no validation as strings, therefore showing error messages for no reason. This was occurring for a select field with integer values.
This was down to async-validator for some reason converting an "empty" object (with no properties) into a string validation rule, presumably as some kind of fallback. For this change we simply make sure empty objects are never returned, instead an empty array is returned, which is treated as "don't do any validation" which is what we want.